+++ R/package-tablexlsx.R
@@ -1,0 +1,8 @@
+#' @keywords internal
+#' @importFrom magrittr "%>%"
+#' @importFrom methods allNames
+#' @importFrom openxlsx addStyle addWorksheet mergeCells openXL saveWorkbook setColWidths setRowHeights writeData writeDataTable
+#' @importFrom stats setNames
+#' @importFrom cli cli_alert_success
+#' @import utils
+"_PACKAGE"